+++ src/constants.js
@@ -8,4 +8,12 @@ export const ASCIIDOCTOR = "asciidoctor";
 export const JEKYLL = "jekyll";
 export const DOCSIFY = "docsify";
 export const ANTORA = "antora";
+export const MDBOOK = "mdbook";
+export const VITEPRESS = "vitepress";
 export const FALLBACK_DOCTOOL = "fallback";
+
+// Known documentation tools themes
+export const SPHINX_ALABASTER = "alabaster";
+export const SPHINX_FURO = "furo";
+export const SPHINX_READTHEDOCS = "readthedocs";
+export const SPHINX_IMMATERIAL = "immaterial";