+++ src/flyout.js
@@ -2,13 +2,18 @@ import { ajv } from "./data-validation";
 import READTHEDOCS_LOGO_WORDMARK from "./images/logo-wordmark-light.svg";
 import READTHEDOCS_LOGO from "./images/logo-light.svg";
 import { library, icon } from "@fortawesome/fontawesome-svg-core";
-import { faCodeBranch, faLanguage } from "@fortawesome/free-solid-svg-icons";
+import {
+  faCodeBranch,
+  faCaretDown,
+  faLanguage,
+} from "@fortawesome/free-solid-svg-icons";
 import { html, nothing, render, LitElement } from "lit";
 import { classMap } from "lit/directives/class-map.js";
 import { default as objectPath } from "object-path";
 
 import styleSheet from "./flyout.css";
 import { AddonBase, addUtmParameters, getLinkWithFilename } from "./utils";
+import { SPHINX, MKDOCS_MATERIAL } from "./constants";
 import {
   EVENT_READTHEDOCS_SEARCH_SHOW,
   EVENT_READTHEDOCS_FLYOUT_HIDE,
@@ -45,6 +50,19 @@ export class FlyoutElement extends LitElement {
       return;
     }
     this.config = config;
+
+    // The "position" is a value that can be defined from the dashboard.
+    // There are two main options: "Default" or a specific value.
+    // When "Default" is used, the value will be grabbed from the HTML element (e.g. explicitly set by the theme author).
+    // In case it's not defined, the value defined in the `constructor` will be used ("bottom-right")
+    const dashboardPosition = objectPath.get(
+      this.config,
+      "addons.flyout.position",
+      null,
+    );
+    if (dashboardPosition) {
+      this.position = dashboardPosition;
+    }
   }
 
   _close() {
@@ -81,23 +99,28 @@ export class FlyoutElement extends LitElement {
       this.config.projects.current.versioning_scheme !==
       "single_version_without_translations"
     ) {
-      version = html`<span
-        >${iconCodeBranch.node[0]} ${this.config.versions.current.slug}</span
-      >`;
+      version = html`<span class="version">
+        ${iconCodeBranch.node[0]} ${this.config.versions.current.slug}
+      </span> `;
     }
+
+    const iconCaretDown = icon(faCaretDown, {
+      classes: ["icon"],
+    });
 
     let translation = nothing;
     if (this.config.projects.translations.length > 0) {
-      translation = html`<span
-        >${iconLanguage.node[0]}
+      translation = html`<span class="language">
+        ${iconLanguage.node[0]}
         ${this.config.projects.current.language.code}</span
-      >`;
+      > `;
     }
 
     return html`
       <header @click="${this._toggleOpen}">
         <img class="logo" src="${this.readthedocsLogo}" alt="Read the Docs" />
         ${translation} ${version}
+        <span class="caret">${iconCaretDown.node[0]}</span>
       </header>
     `;
   }
@@ -303,6 +326,11 @@ export class FlyoutElement extends LitElement {
     `;
   }
 
+  updateCSSClasses() {
+    this.classes = { floating: this.floating, container: true };
+    this.classes[this.position] = true;
+  }
+
   render() {
     // The element doesn't yet have our config, don't render it.
     if (this.config === null) {
@@ -310,11 +338,10 @@ export class FlyoutElement extends LitElement {
       return nothing;
     }
 
-    const classes = { floating: this.floating, container: true };
-    classes[this.position] = true;
+    this.updateCSSClasses();
 
     return html`
-      <div class=${classMap(classes)}>
+      <div class=${classMap(this.classes)}>
         ${this.renderHeader()}
         <main class=${classMap({ closed: !this.opened })}>
           ${this.renderLanguages()} ${this.renderVersions()}