+++ src/notification.js
@@ -213,6 +213,8 @@ export class NotificationElement extends LitElement {
     if (
       this.readingLatestVersion &&
       this.stableVersionAvailable &&
+      this.config.versions.current.slug !==
+        this.config.projects.current.default_version &&
       objectPath.get(this.config, "addons.notifications.show_on_latest", false)
     ) {
       return this.renderLatestVersionWarning();
@@ -240,8 +242,9 @@ export class NotificationElement extends LitElement {
     //  - if the user is reading the "latest" version: shows a notification to warn
     //    the user about reading the latest development version.
     //  - if the user is reading a non-"stable" version: shows a notification to warn
-    //    the user about reading a version that may be old. Except if the reading version
-    //    is the project's default version.
+    //    the user about reading a version that may be old.
+    //  - Both notifications are skipped if the current version is the **default** version,
+    //    as we assume then the author wants the user to be reading that version.
     //
     // This does not cover all the cases where this notification could be useful,
     // but users with different needs should be able to implement their own custom logic.