+++ CHANGELOG.rst
@@ -3,6 +3,34 @@ CHANGELOG
 
 .. The text for the changelog is manually generated for now.
 
+Version v0.26.0
+---------------
+
+:date: Jan 21, 2025
+
+* Search: small updates on look & feel (:pr:`497`)
+
+Version v0.25.0
+---------------
+
+:date: Jan 14, 2025
+
+* EthicalAds: better positioning for auto-detected Antora and mdBook (:pr:`475`)
+* Flyout: allow theme developers and project owners to define the position (:pr:`482`)
+* Flyout: add custom styles based on auto-detected documentation tool (:pr:`473`)
+* Flyout: show a caret next to the version (:pr:`394`)
+* Flyout: use flex grid for better spacing (:pr:`486`)
+* Packages: updated via ncu -u (:pr:`489`)
+
+Version v0.24.0
+---------------
+
+:date: Jan 8, 2025
+
+* Don't show the latest notification if it's the default version (:pr:`480`)
+* EthicalAd: better ad position for Docsify (:pr:`470`)
+* Packages: updated via ncu -u (:pr:`479`)
+
 Version v0.23.2
 ---------------